First initial commit.
This is the firs commit before the release it  contains all relevant changes
happened in the project. We create this branch to reset the project log.
The old master branch it at the archives/old-flow.

The new master branch is managed with lerna, this is a multi package
repository manager. With his tool we can easily manage versioning
pachages that are co dependent, still you want to share useful
functionality with others, so the best of the conquering techniques.
You don't have to compromise if you want your own monolythic system
operatioan in a smart way (chatedral style development) and yet
be able the share chunks of fine grained functioanities to the masses.